Bill Type enum creation

+++ C0751731_MID_TERM_MAD3004/C0751731_MID_TERM_MAD3004/Hydro.swift
@@ -14,7 +14,7 @@ class Hydro : Bill
     var Agency : String?
     var unitsConsumed : Float?
     
-    override init(bId: Int, bDate: Date, bType: String, bTotalAmount: Double)
+    override init(bId: Int, bDate: Date, bType: String, bTotalAmount: Float)
     {
         self.Agency = ""
         self.unitsConsumed = 0
